Use chart wrapper for area and line charts
