--- lib/AreaChart/AreaChart.d.ts
@@ -1,2 +1,0 @@
-/// <reference types="react" />
-export declare const ResourceAreaChart: (props: any) => JSX.Element;
