+++ lib/AreaChart/index.js
@@ -1,2 +1,2 @@
-export { ResourceAreaChart } from './AreaChart';
+export { ResourceAreaChart } from './ResourceAreaChart';
 //# sourceMappingURL=index.js.map
