--- lib/ChartWrapper/AreaChart.js
@@ -1,4 +1,0 @@
-import { chartWrapper } from "./ChartWrapper";
-import { AreaChart } from "recharts";
-export var ResourceAreaChart = chartWrapper(AreaChart);
-//# sourceMappingURL=AreaChart.js.map
