--- lib/ChartWrapper/LineChart.d.ts
@@ -1,2 +1,0 @@
-/// <reference types="react" />
-export declare const ResourceLineChart: (props: any) => JSX.Element;
