--- lib/LineChart/LineChart.js
@@ -1,4 +1,0 @@
-import { chartWrapper } from "../ChartWrapper";
-import { LineChart } from "recharts";
-export var ResourceLineChart = chartWrapper(LineChart);
-//# sourceMappingURL=LineChart.js.map
