+++ lib/LineChart/ResourceLineChart.d.ts
@@ -1,9 +1,7 @@
-import * as React from 'react';
-import { LineChartProps } from 'recharts';
+/// <reference types="react" />
+import { LineChartProps } from "recharts";
 import { ResourceCollection } from 'webpanel-data';
 export interface ResourceLineChartProps extends LineChartProps {
     resourceCollection: ResourceCollection;
 }
-export declare class ResourceLineChart extends React.Component<ResourceLineChartProps> {
-    render(): JSX.Element;
-}
+export declare const ResourceLineChart: (props: any) => JSX.Element;
