+++ lib/LineChart/index.d.ts
@@ -1,1 +1,1 @@
-export { ResourceLineChart } from './LineChart';
+export { ResourceLineChart } from './ResourceLineChart';
