+++ lib/LineChart/index.js
@@ -1,2 +1,2 @@
-export { ResourceLineChart } from './LineChart';
+export { ResourceLineChart } from './ResourceLineChart';
 //# sourceMappingURL=index.js.map
