--- src/AreaChart/AreaChart.tsx
@@ -1,4 +1,0 @@
-import { chartWrapper } from "../ChartWrapper";
-import { AreaChart } from "recharts";
-
-export const ResourceAreaChart = chartWrapper(AreaChart);
