--- src/LineChart/LineChart.tsx
@@ -1,4 +1,0 @@
-import { chartWrapper } from "../ChartWrapper";
-import { LineChart } from "recharts";
-
-export const ResourceLineChart = chartWrapper(LineChart);
