+++ src/LineChart/ResourceLineChart.tsx
@@ -1,30 +1,9 @@
-import * as React from 'react';
-import { LineChart, LineChartProps } from 'recharts';
-import { observer } from 'mobx-react';
+import { chartWrapper } from "../ChartWrapper";
+import { LineChart, LineChartProps } from "recharts";
 import { ResourceCollection } from 'webpanel-data';
-import { Spin } from 'antd';
 
 export interface ResourceLineChartProps extends LineChartProps {
     resourceCollection: ResourceCollection;
 }
 
-@observer
-export class ResourceLineChart extends React.Component<ResourceLineChartProps> {
-    render(){
-        const {
-            resourceCollection: {data, loading},
-            children,
-            ...restProps
-        } = this.props;
-        
-        if (loading) {
-            return <div style={{textAlign: "center", paddingTop: 32, paddingBottom: 32}}><Spin /></div>;
-        }
-
-        return (
-            <LineChart data={data || []} {...restProps}>
-                {children}
-            </LineChart>
-        );
-    }
-}
+export const ResourceLineChart = chartWrapper(LineChart);
