+++ src/LineChart/index.tsx
@@ -1,1 +1,1 @@
-export { ResourceLineChart } from './LineChart';
+export { ResourceLineChart } from './ResourceLineChart';
